+++ mayan/apps/cabinets/tests/test_models.py
@@ -91,7 +91,7 @@ class CabinetTestCase(CabinetTestMixin, BaseTestCase):
         events = self._get_test_events()
         self.assertEqual(events.count(), 1)
 
-        self.assertEqual(events[0].action_object, None)
+        self.assertEqual(events[0].action_object, self._test_cabinet)
         self.assertEqual(events[0].actor, inner_cabinet)
         self.assertEqual(events[0].target, inner_cabinet)
         self.assertEqual(events[0].verb, event_cabinet_created.id)