+++ mayan/apps/cabinets/tests/test_views.py
@@ -207,7 +207,7 @@ class CabinetChildViewTestCase(
         events = self._get_test_events()
         self.assertEqual(events.count(), 1)
 
-        self.assertEqual(events[0].action_object, None)
+        self.assertEqual(events[0].action_object, self._test_cabinet)
         self.assertEqual(events[0].actor, self._test_case_user)
         self.assertEqual(events[0].target, self._test_cabinet)
         self.assertEqual(events[0].verb, event_cabinet_created.id)